Reorganize custom widgets
